create registry module
Put all method registry related stuff to wakepy.core.registry
