+++ tests/unit/test_core/test_method.py
@@ -2,16 +2,8 @@ import re
 
 import pytest
 
-from wakepy.core.method import (
-    Method,
-    MethodDefinitionError,
-    get_method,
-    get_methods,
-    get_methods_for_mode,
-    method_names_to_classes,
-    register_method,
-    select_methods,
-)
+from wakepy.core.method import Method, select_methods
+from wakepy.core.registry import MethodRegistryError, get_method, get_methods
 
 
 def test_overridden_methods_autodiscovery():
@@ -99,99 +91,27 @@ def test_method_has_x_is_not_writeable():
         MethodWithEnter.has_enter = False
 
 
-def test_get_method_which_is_not_yet_defined(monkeypatch):
-    monkeypatch.setattr("wakepy.core.method._method_registry", dict())
-
-    # The method registry is empty so there is no Methods with the name
-    with pytest.raises(
-        KeyError, match=re.escape('No Method with name "Some name" found!')
-    ):
-        get_method("Some name")
-
-
-def test_get_method_working_example(monkeypatch):
+@pytest.mark.usefixtures("empty_method_registry")
+def test_not_possible_to_define_two_methods_with_same_name(testutils, monkeypatch):
     somename = "Some name"
-    # Make the registry empty
-    monkeypatch.setattr("wakepy.core.method._method_registry", dict())
-
-    # Create a method
-    class SomeMethod(Method):
-        name = somename
-
-    # Check that we can retrieve the method
-    method_class = get_method(somename)
-    assert method_class is SomeMethod
-
-
-def test_not_possible_to_define_two_methods_with_same_name(monkeypatch):
-    somename = "Some name"
-    # Make the registry empty
-    monkeypatch.setattr("wakepy.core.method._method_registry", dict())
 
     class SomeMethod(Method):
         name = somename
 
     # It is not possible to define two methods if same name
     with pytest.raises(
-        MethodDefinitionError, match=re.escape('Duplicate Method name "Some name"')
+        MethodRegistryError, match=re.escape('Duplicate Method name "Some name"')
     ):
 
         class SomeMethod(Method):  # noqa:F811
             name = somename
 
-    # sanity check: The monkeypatching works as we expect
-    monkeypatch.setattr("wakepy.core.method._method_registry", dict())
+    testutils.empty_method_registry(monkeypatch)
 
     # Now as the registry is empty it is possible to define method with
     # the same name again
     class SomeMethod(Method):  # noqa:F811
         name = somename
-
-
-@pytest.mark.usefixtures("provide_methods_a_f")
-def test_method_names_to_classes():
-    (A, B, C) = get_methods(["A", "B", "C"])
-
-    # Asking for a list, getting a list
-    assert method_names_to_classes(["A", "B"]) == [A, B]
-    # The order of returned items matches the order of input params
-    assert method_names_to_classes(["C", "B", "A"]) == [C, B, A]
-    assert method_names_to_classes(["B", "A", "C"]) == [B, A, C]
-
-    # Asking a tuple, getting a tuple
-    assert method_names_to_classes(("A", "B")) == (A, B)
-    assert method_names_to_classes(("C", "B", "A")) == (C, B, A)
-
-    # Asking a set, getting a set
-    assert method_names_to_classes({"A", "B"}) == {A, B}
-    assert method_names_to_classes({"C", "B"}) == {C, B}
-
-    # Asking None, getting None
-    assert method_names_to_classes(None) is None
-
-    # Asking something that does not exists will raise KeyError
-    with pytest.raises(KeyError, match=re.escape('No Method with name "foo" found!')):
-        method_names_to_classes(["A", "foo"])
-
-    # Using unsupported type raises TypeError
-    with pytest.raises(TypeError):
-        method_names_to_classes(4123)
-
-
-@pytest.mark.usefixtures("provide_methods_a_f")
-def test_get_methods_for_mode():
-    methods = get_methods(["A", "B", "C", "D", "E", "F"])
-    (MethodA, MethodB, _, MethodD, MethodE, MethodF) = methods
-
-    assert get_methods_for_mode("first_mode") == [
-        MethodB,
-        MethodD,
-        MethodE,
-    ]
-    assert get_methods_for_mode("second_mode") == [
-        MethodA,
-        MethodF,
-    ]
 
 
 @pytest.mark.usefixtures("provide_methods_a_f")
@@ -228,23 +148,6 @@ def test_select_methods():
         ),
     ):
         select_methods(methods, use_only=["B"], omit=["E"])
-
-
-def test_register_method(monkeypatch):
-    # Make the registry empty
-    monkeypatch.setattr("wakepy.core.method._method_registry", dict())
-
-    class MethodA(Method):
-        name = "A"
-
-    assert get_method("A") is MethodA
-
-    # It is possible to register the same method many times without issues.
-    register_method(MethodA)
-    register_method(MethodA)
-    register_method(MethodA)
-
-    assert get_method("A") is MethodA
 
 
 def test_method_defaults():
